I've refactored the code to fold the TextGenerator into the MarkovModel. This involved moving all the methods from TextGenerator.js to MarkovModel.js, deleting the now-empty TextGenerator.js, and updating all the places in the code that were using the old TextGenerator to use the new MarkovModel instead.
